--- src/test/java/com/moneygaming/tests/test1.java
@@ -1,27 +1,0 @@
-package com.moneygaming.tests;
-
-import com.moneygaming.utilities.WebDriverFactory;
-import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-public class test1 {
-
-    @BeforeMethod
-    public void setUp(){
-        WebDriver driver = WebDriverFactory.getDriver("chrome");
-        driver.manage().window().maximize();
-    }
-
-    @Test
-    public void test1(){
-
-    }
-
-    @AfterMethod
-    public void tearDown(){
-
-    }
-
-}
